Closes #2: Disable xdebug in composer update
This took a little bit of hacking, but I am first creating a separate
directory for binary files I need to download. Then I am downloading the
Composer binary if it does not exist so that I can run Composer with
`php -n` which allows me to run PHP without the php.ini file.

The ./src directory is ignored by Git so it will not be committed.